Updated package.json and fixed typo
The project cannot be used automatically using the typical node require method simply by name without the main directive, and each user needs to reference the specific class directly by inspecting the library more closely.  This requires more effort, and may cause problems that are not obvious to everyone.
I have added that line so this project conforms to most others' convention, so that the library can be used more easily, and I also fixed a typo.